+++ src/app/modules/+lists/containers/lists-main-component/lists-main.component.ts
@@ -6,14 +6,15 @@ import { CartService } from '@shared/services/cart.service';
 import { LoadingService } from '@shared/services/loading.service';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { catchError, finalize, forkJoin, of } from 'rxjs';
-import { Article, Category, User } from '@shared/models/cart.models';
+import { Article, Category, List, User } from '@shared/models/cart.models';
 import { ModalService } from '@shared/services/modal.service';
-import { DEFAULT_MODAL_OPTIONS, ModalOptions } from '@shared/models/modal.model';
+import { DEFAULT_MODAL_OPTIONS } from '@shared/models/modal.model';
 import { AddListComponent } from '@modules/+lists/components/add-list/add-list.component';
-import { formatPrice, getCategory } from '@shared/utils/cart.utils';
+import { formatPrice, getCategory, getNewList } from '@shared/utils/cart.utils';
 import { IconEmum } from '@shared/models/icon.models';
 import { SocialService } from '@shared/services/social.service';
 import { TOAST_STATE, ToastService } from '@shared/services/toast.service';
+import { CartOption } from '@shared/components/select/select.component';
 
 @Component({
   selector: 'app-lists-main-component',
@@ -25,10 +26,14 @@ export class ListsMainComponent implements OnInit{
   tableData: TableRow[] = [];
   tableConfig: TableConfig = this.getTableConfig();
   swLoadingFinished = false;
-  private _articles: Article[] = [];
-  private _categories: Category[] = [];
+  listsOptions: CartOption[] = [];
+  selectedList: List | undefined = undefined;
+  articles: Article[] = [];
+  categories: Category[] = [];
+  user!: User;
+
+  private _lists: List[] = [];
   private _destroyRef = inject(DestroyRef);
-  private _user!: User;
 
   ngOnInit(): void {
     this.fetch();
@@ -44,23 +49,32 @@ export class ListsMainComponent implements OnInit{
   ) { }
 
   handleNewList(): void {
-    this.modalService.open(AddListComponent, DEFAULT_MODAL_OPTIONS);
+    this.modalService.open(AddListComponent, {
+      ...DEFAULT_MODAL_OPTIONS,
+      data: { userId: this.user.id },
+      prevenCloseOutside: true,
+    });
+  }
+
+  handleListChange(listId: string): void {
+    this.selectedList = this._lists.find((list: List) => list.id === listId);
   }
 
   private fetch(): void {
     this.getUser();
-
   }
 
   private getUser(): void {
-    this.loading.show();
     this.socialService.user$
-      .pipe(takeUntilDestroyed(this._destroyRef),
-        finalize(() => this.loading.hide()))
+      .pipe(takeUntilDestroyed(this._destroyRef))
       .subscribe({
         next: (user) => {
-          this._user = user;
-          this.loadData();
+          if(user) {
+            this.user = user;
+            this.loadData();
+          } else {
+            this.toast.showToast(TOAST_STATE.ERROR, this.translate.instant('TOAST.USER'));
+          }
         },
         error: () => this.toast.showToast(TOAST_STATE.ERROR, this.translate.instant('TOAST.USER')),
       });
@@ -77,25 +91,64 @@ export class ListsMainComponent implements OnInit{
         .pipe(catchError(() => of([])),
           takeUntilDestroyed(this._destroyRef)
         ),
+      this.cartService.getListsByUser(this.user.email)
+        .pipe(catchError(() => of([])),
+          takeUntilDestroyed(this._destroyRef)
+        ),
     ])
       .pipe(finalize(() => {
         this.loading.hide();
         this.swLoadingFinished = true;
       }))
-      .subscribe(([articles, categories]: [Article[], Category[]]) => {
-        this._articles = articles;
-        this._categories = categories;
-        this.tableData = this._articles.map((article: Article) => ({
+      .subscribe(([articles, categories, lists]: [Article[], Category[], List[]]) => {
+        this.articles = articles;
+        this.categories = categories;
+        this._lists = lists;
+        this.tableData = this.articles.map((article: Article) => ({
           ...article,
           averagePrice: formatPrice(article.averagePrice),
-          category: getCategory(article, this._categories),
+          category: getCategory(article, this.categories),
         }));
         this.tableConfig = this.getTableConfig();
         if (articles.length === NUMBERS.N_0) {
           this.toast.showToast(TOAST_STATE.ERROR, this.translate.instant('TOAST.ARTICLES_OR_CATEGORIES'));
         }
+        this.handleLists(lists);
       })
   }
+
+  private handleLists(lists: List[]): void {
+    if (lists.length === NUMBERS.N_0) {
+      this.createInitialList();
+      return;
+    }
+    this.listsOptions = this.getListOptions(lists);
+    this.selectedList = lists[NUMBERS.N_0];
+  }
+
+  private createInitialList(): void {
+    const list: List = getNewList(this.translate, this.user.id);
+    // this.loading.show();
+    // this.cartService.postList(list)
+    //   .pipe(takeUntilDestroyed(this._destroyRef),
+    //     finalize(() => this.loading.hide()))
+    //   .subscribe({
+    //     next: () => {
+          this._lists.push(list);
+          this.listsOptions = this.getListOptions(this._lists);
+          this.selectedList = list;
+      //   },
+      //   error: () => this.toast.showToast(TOAST_STATE.ERROR, this.translate.instant('TOAST.LIST')),
+      // });
+  }
+
+  private getListOptions(lists: List[]): CartOption[] {
+    return lists.map((list: List) => ({
+      value: list.id,
+      label: list.name,
+    }));
+  }
+
 
   private getTableConfig(): TableConfig {
     const literals = this.translate.instant('LISTS.HEADERS');