+++ src/app/shared/constants/style.constants.ts
@@ -1,0 +1,7 @@
+export const BUTTON_CLASS = {
+  PRIMARY: 'primary-button',
+  SECONDARY: 'secondary-button',
+  LINK: 'link-button',
+  ICON: 'icon-button',
+  CARD_ICON: 'card-icon-button',
+}