+++ src/app/shared/utils/cart.utils.ts
@@ -1,6 +1,8 @@
-import { Article, Category } from "@shared/models/cart.models";
+import { Article, Category, List } from "@shared/models/cart.models";
 import { stringFrom } from "./string.utils";
 import { NUMBERS } from "@shared/constants/number.constants";
+import { STRING_EMPTY } from "@shared/constants/string.constants";
+import { TranslateService } from "@ngx-translate/core";
 
 export const getCategory = (article: Article, categories: Category[]): string => {
   const category = categories.find((c: Category) => article.categories?.includes(c.id));
@@ -10,3 +12,13 @@ export const getCategory = (article: Article, categories: Category[]): string =>
 export const formatPrice = (price: number): string => {
   return price ? `${price.toFixed(NUMBERS.N_2)} €` : '0';
 }
+
+export const getNewList = (translate: TranslateService, userId: string): List => ({
+  id: STRING_EMPTY,
+  name: translate.instant('LISTS.NEW_LIST'),
+  articleList: [],
+  totalPrize: NUMBERS.N_0,
+  userId,
+  creationDate: new Date().toISOString(),
+  isPublic: false
+});